+++ src/datasources/work-items/utils.test.ts
@@ -1,7 +1,7 @@
 import { CUSTOM_PROPERTY_SUFFIX, TAKE_LIMIT } from './constants';
 import { takeErrorMessages } from './constants/QueryEditor.constants';
 import { WorkItemPropertiesOptions, WorkItemTypeOptions } from './types';
-import { getTakeError, isPropertiesNonEmpty, isTakeValid, isTypesNonEmpty, stripCustomPropertySuffix, transformDuration } from './utils';
+import { getTakeError, isPropertiesNonEmpty, isTakeValid, isTypesNonEmpty, stripCustomPropertySuffix } from './utils';
 
 describe('getTakeError', () => {
   it('should return no error for a value within the valid range', () => {
@@ -89,10 +89,4 @@ describe('isTakeValid', () => {
   });
 });
 
-test('transformDuration', () => {
-    expect(transformDuration(0)).toBe('0 sec');
-    expect(transformDuration(61)).toBe('1 min, 1 sec');
-    expect(transformDuration(3661)).toBe('1 hr, 1 min, 1 sec');
-    expect(transformDuration(90061)).toBe('1 day, 1 hr, 1 min, 1 sec');
-    expect(transformDuration(172800)).toBe('2 days');
-});
+