+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [5.18.2](https://github.com/ni/systemlink-grafana-plugins/compare/v5.18.1...v5.18.2) (2026-09-22)
+
+### Bug Fixes
+
+* **work-items:** remove target location and parent columns for fixtures ([#773](https://github.com/ni/systemlink-grafana-plugins/issues/773)) ([94348b9](https://github.com/ni/systemlink-grafana-plugins/commit/94348b901c5c956da613cf10f6821c1f59bf2ae7))
+
+## [5.18.1](https://github.com/ni/systemlink-grafana-plugins/compare/v5.18.0...v5.18.1) (2026-09-21)
+
+### Bug Fixes
+
+* **work-items:** update filter operations to use new 'is empty' logic ([#769](https://github.com/ni/systemlink-grafana-plugins/issues/769)) ([b30cdc1](https://github.com/ni/systemlink-grafana-plugins/commit/b30cdc1ed58c72bbefc8067a2f41ab18e38ec9c6))
+
 ## [5.18.0](https://github.com/ni/systemlink-grafana-plugins/compare/v5.17.0...v5.18.0) (2026-09-18)
 
 ### Features